--- mentionserver/src/main/java/com.mention/Workflow/StringMakerOne.java
@@ -1,8 +1,0 @@
-package com.mention.Workflow;
-
-public class StringMakerOne {
-    public String getStringOne() {
-        String firstString = "He often seeks a lordly dish";
-        return firstString;
-    }
-}
