--- mentionserver/src/main/java/com.mention/Workflow/StringMakerTwo.java
@@ -1,11 +1,0 @@
-package com.mention.Workflow;
-
-public class StringMakerTwo {
-    public String getStringTwo(){
-        String secondString = "to serve upon his table.";
-        return secondString;
-    }
-}
-
-
-
