+++ mentionserver/src/main/java/com.mention/workflow/StringMakerOne.java
@@ -1,0 +1,8 @@
+package com.mention.workflow;
+
+public class StringMakerOne {
+  public String getStringOne() {
+    String firstString = "He often seeks a lordly dish";
+    return firstString;
+  }
+}
