+++ mentionserver/src/main/java/com.mention/workflow/StringMakerTwo.java
@@ -1,0 +1,11 @@
+package com.mention.workflow;
+
+public class StringMakerTwo {
+  public String getStringTwo() {
+    String secondString = "to serve upon his table.";
+    return secondString;
+  }
+}
+
+
+
